Only remove old JRuby jar if it's present

+++ config/emr-bootstrap/remove_old_jruby.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-rm /home/hadoop/lib/jruby-complete-no-joda-1.6.5.jar
+if [ -e /home/hadoop/lib/jruby-complete-no-joda-1.6.5.jar ]
+  then
+    rm /home/hadoop/lib/jruby-complete-no-joda-1.6.5.jar
+fi
